backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
google-genai>=1.0.0
requests>=2.31.0
